+++ packages/editor-core/src/editor/types/index.ts
@@ -1,5 +1,0 @@
-export * from './editor-instance';
-export * from './editor-config';
-export * from './editor-plugin';
-export * from './editor-props';
-export * from './editor-appearance-component';