+++ packages/navigation/src/components/styled/ContainerLogo.jsx
@@ -2,8 +2,8 @@ import styled from 'styled-components';
 import { gridSize } from '../../shared-variables';
 
 const ContainerLogo = styled.div`
-  height: ${gridSize * 5}px;
-  padding: ${gridSize}px ${gridSize / 2}px;
+  height: ${gridSize * 4}px;
+  padding: ${gridSize * 1.5}px ${gridSize / 2}px;
 `;
 
 ContainerLogo.displayName = 'ContainerLogo';