+++ packages/xflow/src/request-or-start-trial/index.jsx
@@ -1,3 +1,4 @@
 export { default as StartTrial } from './components/StartTrial';
+export { default as ContextualStartTrial } from './components/ContextualStartTrial';
 export { default as RequestTrial } from './components/RequestTrial';
 export { default as AlreadyStarted } from './components/AlreadyStarted';