+++ packages/xflow/src/request-or-start-trial/styled/ContextualConfirmTrialContent.jsx
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+import { gridSize, math } from '@atlaskit/theme';
+
+const ContextualConfirmTrialContent = styled.div`
+  text-align: center;
+  margin: auto;
+  max-width: ${math.multiply(gridSize, 40)}px;
+  
+  button {
+    margin-top: ${math.multiply(gridSize, 3)}px;
+  }
+`;
+
+ContextualConfirmTrialContent.displayName = 'ContextualConfirmTrialContent';
+export default ContextualConfirmTrialContent;