+++ packages/dropdown-menu/docs/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/dropdown-menu
 
+## 3.11.1 (2017-11-02)
+
+* bug fix; added missing dependencies (issues closed: ak-3782) ([4dbc3ef](https://bitbucket.org/atlassian/atlaskit/commits/4dbc3ef))
 ## 3.11.0 (2017-10-27)
 
 * feature; use shared HOC from item ([f966d9c](https://bitbucket.org/atlassian/atlaskit/commits/f966d9c))